1:none

+++ src/net/DataSocket.h
@@ -40,7 +40,7 @@ public:
     typedef std::shared_ptr<std::string>                                        PACKET_PTR;
 
 public:
-    explicit DataSocket(sock fd, int maxRecvBufferSize);
+    explicit DataSocket(sock fd, size_t maxRecvBufferSize);
     ~DataSocket();
 
     /*  仅在网络线程中调用才可能返回成功 */
